+++ cm/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.2.1] — 2025-03-16
 
 ### Fixed
 
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 Initial version, extracted into module [`go.bytecodealliance.org/cm`](https://pkg.go.dev/go.bytecodealliance.org/cm).
 
-[Unreleased]: <https://github.com/bytecodealliance/go-modules/compare/cm/v0.2.0..HEAD>
+[Unreleased]: <https://github.com/bytecodealliance/go-modules/compare/cm/v0.2.1..HEAD>
+[v0.2.1]: <https://github.com/bytecodealliance/go-modules/compare/cm/v0.2.0..cm/v0.2.1>
 [v0.2.0]: <https://github.com/bytecodealliance/go-modules/compare/cm/v0.1.0..cm/v0.2.0>
 [v0.1.0]: <https://github.com/bytecodealliance/go-modules/tree/cm/v0.1.0>